requirements.txt
affine==2.3.0
albumentations==0.4.3
asttokens==2.0.3
attrs==19.3.0
bcrypt==3.1.7
buzzard==0.6.4
certifi==2019.11.28
cffi==1.14.0
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
colorama==0.4.3
cryptography==2.8
cycler==0.10.0
decorator==4.4.1
executing==0.4.2
GDAL==2.4.4
GPUtil==1.4.0
icecream==2.0.0
imageio==2.6.1
imgaug==0.2.6
joblib==0.14.1
kiwisolver==1.1.0
matplotlib==3.1.3
networkx==2.4
numpy==1.18.1
opencv-python==4.2.0.32
pandas==1.0.1
paramiko==2.7.1
Pillow==7.0.0
Pint==0.10.1
pycparser==2.19
Pygments==2.5.2
PyNaCl==1.3.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
rasterio==1.1.2
Rtree==0.9.4
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.4.1
Shapely==1.7.0
six==1.14.0
snuggs==1.4.7
sortedcontainers==2.1.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.1
